add note to scrollview spec
found a way tests can false negative— if their imageviews are "out of order" in their content view, the tests will not pass because they are expecting them to be sequentially shown. For example, if contentView.subviews[4] is actually the first image you see on screen, this will fail regardless. Only alternative would be to use accessibility labels, would have to include them in the README.